90% of news functionality complete
