+++ client/src/components/pages/mainPages/NewsPageComponents/EditButton.tsx
@@ -9,7 +9,7 @@ import {
   DialogClose,
   Dialog,
 } from "@/components/ui/dialog";
-import { SquarePen } from "lucide-react";
+import { SquarePen, X } from "lucide-react";
 import { useSettings } from "../../profileDependents/settings/settingsLogic/SettingsContext";
 import { FieldGroup, Field } from "@/components/ui/field";
 import { Input } from "@/components/ui/input";
@@ -18,8 +18,8 @@ import type { NewsPost } from "@/types/PageTypes";
 import ReactMarkdown from "react-markdown";
 import remarkGfm from "remark-gfm";
 import { RadioGroupChoiceCard } from "./RadioGroupChoiceCard";
-import { ResizableVertical } from "./ResizeableVertical";
-import { ResizableHorizontal } from "./ResizeableHorizontal";
+import { ResizableVertical } from "./ResizableVertical";
+import { ResizableHorizontal } from "./ResizableHorizontal";
 
 export function EditButton({
   post,
@@ -117,17 +117,36 @@ export function EditButton({
                   ref={fileInputRef}
                   onChange={handleFileChange}
                 />
+                {imageAlt && (
+                  <div className="flex items-center gap-1">
+                    <span className="text-xs text-muted-foreground truncate">
+                      Current: {imageAlt}
+                    </span>
+                    <button
+                      type="button"
+                      className="shrink-0 cursor-pointer rounded-md p-0.5 hover:bg-muted hover:text-red-500"
+                      onClick={() => {
+                        setImage("");
+                        setImageAlt("");
+                        if (fileInputRef.current)
+                          fileInputRef.current.value = "";
+                      }}
+                    >
+                      <X className="h-3.5 w-3.5" />
+                    </button>
+                  </div>
+                )}
               </div>
               <div className="ml-auto">
                 {imagePosition === "Top" ? (
                   <ResizableVertical
                     onImageSizeChange={setImageSize}
-                    initialImageSize={imageSize}
+                    initialImageSize={post.imageSize ?? 25}
                   />
                 ) : (
                   <ResizableHorizontal
                     onImageSizeChange={setImageSize}
-                    initialImageSize={imageSize}
+                    initialImageSize={post.imageSize ?? 25}
                   />
                 )}
               </div>
