--- client/src/components/pages/mainPages/NewsPageComponents/ResizeableHorizontal.tsx
@@ -1,37 +1,0 @@
-import {
-  ResizableHandle,
-  ResizablePanel,
-  ResizablePanelGroup,
-} from "@/components/ui/resizable";
-
-export function ResizableHorizontal({
-  onImageSizeChange,
-  initialImageSize = 25,
-}: {
-  onImageSizeChange?: (size: number) => void;
-  initialImageSize?: number;
-}) {
-  return (
-    <div onPointerDownCapture={(e) => e.stopPropagation()}>
-      <ResizablePanelGroup
-        orientation="horizontal"
-        className="min-h-50 max-w-sm rounded-lg border md:min-w-112.5"
-      >
-        <ResizablePanel defaultSize={100 - initialImageSize}>
-          <div className="flex h-full items-center justify-center p-6">
-            <span className="font-semibold">Content</span>
-          </div>
-        </ResizablePanel>
-        <ResizableHandle withHandle />
-        <ResizablePanel
-          defaultSize={initialImageSize}
-          onResize={(size) => onImageSizeChange?.(size.asPercentage)}
-        >
-          <div className="flex h-full items-center justify-center p-6">
-            <span className="font-semibold">Image</span>
-          </div>
-        </ResizablePanel>
-      </ResizablePanelGroup>
-    </div>
-  );
-}
